+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/STPAPIClient+AutocompleteTest.swift
@@ -12,19 +12,33 @@ import XCTest
 @testable @_spi(STP) import StripePaymentSheet
 
 class STPAPIClientAutocompleteTest: STPNetworkStubbingTestCase {
-    func _testAutocomplete() async throws {
-        let response = try await makeAPIClient().autocomplete(
+    func testAutocompleteAndDetails() async throws {
+        let sessionToken = UUID().uuidString
+        let locale = "en-US"
+        let autocompleteResponse = try await makeAPIClient().getAddressSuggestions(
             searchText: "354 Oyster Point",
-            locale: "en-US",
+            locale: locale,
             countryCodes: ["US"],
-            sessionToken: UUID().uuidString
+            sessionToken: sessionToken
         )
-        XCTAssertFalse(response.source.isEmpty)
-        XCTAssertFalse(response.suggestions.isEmpty)
-        let first = try XCTUnwrap(response.suggestions.first)
+        XCTAssertFalse(autocompleteResponse.source.isEmpty)
+        XCTAssertFalse(autocompleteResponse.suggestions.isEmpty)
+        let first = try XCTUnwrap(autocompleteResponse.suggestions.first)
         XCTAssertFalse(first.title.isEmpty)
         XCTAssertFalse(first.subtitle.isEmpty)
         XCTAssertFalse(first.matches.isEmpty)
+        if let placeId = first.placeId {
+            let detailsResponse = try await makeAPIClient().getAddressDetails(
+                placeId: placeId,
+                source: autocompleteResponse.source,
+                locale: locale,
+                displayTitle: first.title,
+                sessionToken: sessionToken
+            )
+            XCTAssertNotNil(detailsResponse.address.line1)
+        } else {
+            XCTAssertNotNil(first.address)
+        }
     }
 }